+++ Editor/AssemblyInfo.cs
@@ -1,2 +1,3 @@
 using System.Runtime.CompilerServices;
-[assembly: InternalsVisibleTo("Unity.2D.Common.EditorTests")]
+[assembly: InternalsVisibleTo("Unity.2D.Common.Tests.EditorTests")]
+[assembly: InternalsVisibleTo("Unity.2D.PsdImporter.Editor")]