+++ Editor/ImagePacker/Jobs/FindTightRectJob.cs
@@ -4,7 +4,7 @@ using Unity.Collections;
 using Unity.Collections.LowLevel.Unsafe;
 using Unity.Jobs;
 
-namespace UnityEditor.Experimental.U2D.Common
+namespace UnityEditor.U2D.Common
 {
     public struct FindTightRectJob : IJobParallelFor
     {